--- SNC-FE/src/app/models/question.model.ts
@@ -1,5 +1,0 @@
-export class Question {
-    questiontext: string;
-    answertype: string;
-    _id: string;
-}
