--- SNC-FE/src/app/models/saqtemplate.models.ts
@@ -1,7 +1,0 @@
-import {Question} from "./question.model";
-
-export class SAQTemplate {
-    _id: string;
-    name: string;
-    questions: [Question];
-}
