--- SNC-FE/src/app/models/subUser.model.ts
@@ -1,9 +1,0 @@
-export class SubUser {
-    _id: string;
-    email: string;
-    password: string;
-    fname: string;
-    lname: string;
-    telephone: string;
-    superuserid: string;
-}
