--- SNC-FE/src/app/models/superUser.model.ts
@@ -1,13 +1,0 @@
-import {SubUser} from "./subUser.model";
-
-export class SuperUser {
-    _id: string;
-    email: string;
-    password: string;
-    fname: string;
-    lname: string;
-    address: string;
-    company: string;
-    telephone: string;
-    subusers: [SubUser]
-}
