+++ SNC-FE/src/components/login/login.component.html
@@ -10,10 +10,20 @@
         </h2>
         <div class="form-group card-body text-center justify-content-center">
 
-          Username<input type="text" formControlName="username" class="form-control form-control-lg" />
+        <label for="username">Username</label>
+        <input type="text" formControlName="username" class="form-control form-control-lg" [ngClass]="{ 'is-invalid': submitted && form.username.errors }"/>
+        <div *ngIf="submitted && form.username.errors" class="invalid-feedback">
+            <div *ngIf="form.username.errors.required">Username is required</div>
+        </div>
 
-          Password<input type="password" formControlName="password" class="form-control form-control-lg" />
 
+
+          <label for="password">Password</label>
+          <input type="password" formControlName="password" class="form-control form-control-lg" [ngClass]="{ 'is-invalid': submitted && form.password.errors }"  />
+          <div *ngIf="submitted && form.password.errors" class="invalid-feedback">
+            <div *ngIf="form.password.errors.required">Password is required</div>
+            <div *ngIf="form.password.errors.minlength">Password must be at least 8 characters</div>
+          </div>
 
         </div>
