+++ SNC-FE/src/components/register/register.component.html
@@ -11,44 +11,44 @@
         </h2>
         <div class="form-group card-body text-center justify-content-center">
 
-          <input type="text" formControlName="first" placeholder="First Name"  class="form-control form-control-lg" />
-          <div *ngIf="f.first.errors" class="invalid-feedback">
-            <div *ngIf="f.first.errors.required">Answer is required</div>
+          <input type="text" formControlName="first" placeholder="First Name"  class="form-control form-control-lg"  [ngClass]="{ 'is-invalid': submitted && form.first.errors }" />
+          <div *ngIf="form.first.errors" class="invalid-feedback">
+            <div *ngIf="form.first.errors.required">Answer is required</div>
           </div>
 
-          <input type="text" formControlName="last" placeholder="Last Name" class="form-control form-control-lg" />
-          <div *ngIf="f.last.errors" class="invalid-feedback">
-            <div *ngIf="f.last.errors.required">Answer is required</div>
+          <input type="text" formControlName="last" placeholder="Last Name" class="form-control form-control-lg" [ngClass]="{ 'is-invalid': submitted && form.last.errors } "/>
+          <div *ngIf="form.last.errors" class="invalid-feedback">
+            <div *ngIf="form.last.errors.required">Answer is required</div>
           </div>
 
-          <input type="text" formControlName="email" placeholder="Email Address" class="form-control form-control-lg" />
-          <div *ngIf="f.email.errors" class="invalid-feedback">
-            <div *ngIf="f.email.errors.required">Answer is required</div>
+          <input type="text" formControlName="email" placeholder="Email Address" class="form-control form-control-lg"  [ngClass]="{ 'is-invalid': submitted && form.email.errors }" />
+          <div *ngIf="form.email.errors" class="invalid-feedback">
+            <div *ngIf="form.email.errors.required">Answer is required</div>
           </div>
 
-          <input type="text" formControlName="address" placeholder="Address" class="form-control form-control-lg" />
-          <div *ngIf="f.address.errors" class="invalid-feedback">
-            <div *ngIf="f.address.errors.required">Answer is required</div>
+          <input type="text" formControlName="address" placeholder="Address" class="form-control form-control-lg"  [ngClass]="{ 'is-invalid': submitted && form.address.errors }"/>
+          <div *ngIf="form.address.errors" class="invalid-feedback">
+            <div *ngIf="form.address.errors.required">Answer is required</div>
           </div>
 
-          <input type="text" formControlName="company" placeholder="Company Name" class="form-control form-control-lg" />
-          <div *ngIf="f.company.errors" class="invalid-feedback">
-            <div *ngIf="f.company.errors.required">Answer is required</div>
+          <input type="text" formControlName="company" placeholder="Company Name" class="form-control form-control-lg"  [ngClass]="{ 'is-invalid': submitted && form.company.errors }"/>
+          <div *ngIf="form.company.errors" class="invalid-feedback">
+            <div *ngIf="form.company.errors.required">Answer is required</div>
           </div>
 
-          <input type="text" formControlName="phone" placeholder="First Name"  class="form-control form-control-lg" />
-          <div *ngIf="f.phone.errors" class="invalid-feedback">
-            <div *ngIf="f.phone.errors.required">Answer is required</div>
+          <input type="text" formControlName="phone" placeholder="Phone Number"  class="form-control form-control-lg"  [ngClass]="{ 'is-invalid': submitted && form.phone.errors }"/>
+          <div *ngIf="form.phone.errors" class="invalid-feedback">
+            <div *ngIf="form.phone.errors.required">Answer is required</div>
           </div>
 
-          <input type="text" formControlName="password" placeholder="Password" class="form-control form-control-lg" />
-          <div *ngIf="f.password.errors" class="invalid-feedback">
-            <div *ngIf="f.password.errors.required">Answer is required</div>
+          <input type="text" formControlName="password" placeholder="Password" class="form-control form-control-lg"  [ngClass]="{ 'is-invalid': submitted && form.password.errors }"/>
+          <div *ngIf="form.password.errors" class="invalid-feedback">
+            <div *ngIf="form.password.errors.required">Answer is required</div>
           </div>
 
-          <input type="text" formControlName="passConf" placeholder="Confirm Password" class="form-control form-control-lg" />
-          <div *ngIf="f.passConf.errors" class="invalid-feedback">
-            <div *ngIf="f.passConf.errors.required">Answer is required</div>
+          <input type="text" formControlName="passConf" placeholder="Confirm Password" class="form-control form-control-lg" [ngClass]="{ 'is-invalid': submitted && form.passConf.errors }" />
+          <div *ngIf="form.passConf.errors" class="invalid-feedback">
+            <div *ngIf="form.passConf.errors.required">Answer is required</div>
           </div>
 
         </div>
