+++ SNC-FE/src/models/question.model.ts
@@ -1,0 +1,5 @@
+export class Question {
+    questiontext: string;
+    answertype: string;
+    _id: string;
+}
