+++ SNC-FE/src/models/saqtemplate.models.ts
@@ -1,0 +1,7 @@
+import {Question} from "./question.model";
+
+export class SAQTemplate {
+    _id: string;
+    name: string;
+    questions: [Question];
+}
