+++ SNC-FE/src/models/subUser.model.ts
@@ -1,0 +1,9 @@
+export class SubUser {
+    _id: string;
+    email: string;
+    password: string;
+    fname: string;
+    lname: string;
+    telephone: string;
+    superuserid: string;
+}
