+++ SNC-FE/src/models/superUser.model.ts
@@ -1,0 +1,13 @@
+import {SubUser} from "./subUser.model";
+
+export class SuperUser {
+    _id: string;
+    email: string;
+    password: string;
+    fname: string;
+    lname: string;
+    address: string;
+    company: string;
+    telephone: string;
+    subusers: [SubUser]
+}
